Use non alpha-premultiplied components
Calls to color.Color.RGBA() returns alpha-premultiplied components.
However, the hue changer algorithm implemented in palette package
considers that the r, g, b input components are in the
NO-alpha-premultiplied status (values from 0 to 255).

So for the algorithm to generate the correct values, divide
the input components by 0x101 to scale them to the proper values.

This commit fix #1